add abstraction of services at the interface level
